MOOOOOOAR ABSTRACTION
