+++ EppoEngine/Source/Core/Base.h
@@ -38,18 +38,9 @@ namespace Eppo
 	{
 		return std::make_unique<T>(std::forward<Args>(args)...);
 	}
-
-	// Shared pointer wrapper
-	template<typename T>
-	using Ref = std::shared_ptr<T>;
-
-	template<typename T, typename... Args>
-	constexpr Ref<T> CreateRef(Args&&... args)
-	{
-		return std::make_shared<T>(std::forward<Args>(args)...);
-	}
 }
 
+#include "Core/Ref.h"
 #include "Core/Assert.h"
 #include "Core/Log.h"
 #include <Debug/Profiler.h>
