+++ EppoEngine/Source/Core/Log.cpp
@@ -5,7 +5,7 @@
 
 namespace Eppo
 {
-	Ref<spdlog::logger> Log::s_CoreLogger;
+	std::shared_ptr<spdlog::logger> Log::s_CoreLogger;
 
 	void Log::Init()
 	{
