+++ EppoEngine/Source/Core/Log.h
@@ -12,10 +12,10 @@ namespace Eppo
 	public:
 		static void Init();
 
-		static Ref<spdlog::logger>& GetCoreLogger() { return s_CoreLogger; }
+		static std::shared_ptr<spdlog::logger>& GetCoreLogger() { return s_CoreLogger; }
 
 	private:
-		static Ref<spdlog::logger> s_CoreLogger;
+		static std::shared_ptr<spdlog::logger> s_CoreLogger;
 	};
 }
