+++ EppoEngine/Source/Platform/Vulkan/Descriptor/DescriptorLayoutInfo.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "Renderer/Vulkan.h"
+#include "Platform/Vulkan/Vulkan.h"
 
 #include <vector>
