+++ EppoEngine/Source/Platform/Vulkan/VulkanRenderer.h
@@ -1,0 +1,18 @@
+#pragma once
+
+#include "Renderer/Renderer.h"
+
+namespace Eppo
+{
+	class VulkanRenderer : public Renderer
+	{
+	public:
+		VulkanRenderer() = default;
+		~VulkanRenderer() = default;
+
+		void Init() override;
+		void Shutdown() override;
+
+	private:
+	};
+}
