+++ EppoEngine/Source/Platform/Vulkan/VulkanVertexBuffer.h
@@ -1,0 +1,29 @@
+#pragma once
+
+#include "Core/Buffer.h"
+#include "Platform/Vulkan/Allocator.h"
+#include "Renderer/Buffer/VertexBuffer.h"
+
+namespace Eppo
+{
+	class VulkanVertexBuffer : public VertexBuffer
+	{
+	public:
+		VulkanVertexBuffer(void* data, uint32_t size);
+		VulkanVertexBuffer(uint32_t size);
+		~VulkanVertexBuffer();
+
+		void SetData(void* data, uint32_t size) override;
+
+		VkBuffer GetBuffer() const { return m_Buffer; }
+
+	private:
+		void CreateBuffer(VmaMemoryUsage usage);
+
+	private:
+		Buffer m_LocalStorage;
+
+		VkBuffer m_Buffer;
+		VmaAllocation m_Allocation;
+	};
+}
