+++ EppoEngine/Source/Renderer/Buffer/VertexBuffer.h
@@ -1,30 +1,15 @@
 #pragma once
 
-#include "Core/Buffer.h"
-#include "Renderer/Allocator.h"
-
 namespace Eppo
 {
 	class VertexBuffer
 	{
 	public:
-		VertexBuffer(void* data, uint32_t size);
-		VertexBuffer(uint32_t size);
-		~VertexBuffer();
-		VertexBuffer(const VertexBuffer&) = delete;
-		VertexBuffer& operator=(const VertexBuffer&) = delete;
+		virtual ~VertexBuffer() {};
 
-		void SetData(void* data, uint32_t size);
+		virtual void SetData(void* data, uint32_t size) = 0;
 
-		VkBuffer GetBuffer() const { return m_Buffer; }
-
-	private:
-		void CreateBuffer(VmaMemoryUsage usage);
-
-	private:
-		Buffer m_LocalStorage;
-
-		VkBuffer m_Buffer;
-		VmaAllocation m_Allocation;
+		static Ref<VertexBuffer> Create(void* data, uint32_t size);
+		static Ref<VertexBuffer> Create(uint32_t size);
 	};
 }
