+++ EppoEngine/Source/Renderer/RendererContext.h
@@ -10,7 +10,7 @@ namespace Eppo
 		Vulkan
 	};
 
-	class RendererContext
+	class RendererContext : public RefCounter
 	{
 	public:
 		virtual ~RendererContext() {}
@@ -18,21 +18,11 @@ namespace Eppo
 		virtual void Init() = 0;
 		virtual void Shutdown() = 0;
 
-		void WaitIdle();
-
-		GLFWwindow* GetWindowHandle() { return m_WindowHandle; }
-
-		TracyVkCtx GetCurrentProfilerContext() { return m_TracyContexts[m_Swapchain->GetCurrentImageIndex()]; }
-
 		static RendererAPI GetAPI() { return s_API; }
-		static Scope<RendererContext> Create(void* windowHandle);
+		static Ref<RendererContext> Get();
+		static Ref<RendererContext> Create(void* windowHandle);
 
 	private:
 		static RendererAPI s_API;
-
-		GLFWwindow* m_WindowHandle = nullptr;
-
-		// Tracy profiler context
-		std::vector<TracyVkCtx> m_TracyContexts;
 	};
 }
